On Windows, allow mmaps at file offsets > 0.
There were two problems. The first was that CreateFileMapping
was being called with 0 as the high-order DWORD for the maximum
size. Now the high and low order DWORDs are computed and passed
correctly.

Secondly, the maximum size passed to CreateFileMapping is not the
size of the mmap'ed region (the "view" of the file). Instead, it
is the maximum size of the file from which mapped views can be created.
Thus, maximum size in this context is the file offset + the requested
length.